@@ -11,15 +11,15 @@
         }
     ],
     "require": {
-        "php": "^5.3.9|^7.0",
+        "php": "^7.0",
         "puli/repository": "^1.0-beta11@dev,<1.1",
         "puli/discovery": "^1.0.0-beta10@dev",
         "webmozart/glob": "^4.0",
         "webmozart/path-util": "^2.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.6",
-        "sebastian/version": "^1.0.1"
+        "phpunit/phpunit": "^6.5",
+        "sebastian/version": "^2.0"
     },
     "autoload": {
         "psr-4": {
